refactor : register, login -> Intro로 합치기
