+++ app/src/main/java/com/surveasy/surveasy/presentation/intro/HelperText.kt
@@ -1,4 +1,4 @@
-package com.surveasy.surveasy.presentation.register
+package com.surveasy.surveasy.presentation.intro
 
 enum class HelperText {
     NONE,
