+++ app/src/main/java/com/surveasy/surveasy/presentation/intro/IntroActivity.kt
@@ -1,11 +1,6 @@
-package com.surveasy.surveasy.presentation.register
+package com.surveasy.surveasy.presentation.intro
 
-import android.app.Activity
-import android.content.Context
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.util.AttributeSet
-import android.view.View
 import androidx.activity.viewModels
 import androidx.navigation.NavController
 import androidx.navigation.fragment.NavHostFragment
@@ -13,7 +8,7 @@ import com.surveasy.surveasy.R
 import com.surveasy.surveasy.databinding.ActivityRegisterBinding
 import com.surveasy.surveasy.presentation.base.BaseActivity
 
-class RegisterActivity : BaseActivity<ActivityRegisterBinding>(ActivityRegisterBinding::inflate) {
+class IntroActivity : BaseActivity<ActivityRegisterBinding>(ActivityRegisterBinding::inflate) {
     private val vm: RegisterViewModel by viewModels()
     private var navController: NavController? = null
     override fun onCreate(savedInstanceState: Bundle?) {
