+++ app/src/main/java/com/surveasy/surveasy/presentation/intro/RegisterAgreeFragment.kt
@@ -1,20 +1,10 @@
-package com.surveasy.surveasy.presentation.register
+package com.surveasy.surveasy.presentation.intro
 
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.CompoundButton
-import android.widget.CompoundButton.OnCheckedChangeListener
-import androidx.fragment.app.activityViewModels
 import androidx.fragment.app.viewModels
-import androidx.navigation.NavController
-import androidx.navigation.Navigation
 import androidx.navigation.fragment.findNavController
 import com.surveasy.surveasy.R
 import com.surveasy.surveasy.databinding.FragmentRegisterAgreeBinding
 import com.surveasy.surveasy.presentation.base.BaseFragment
-import kotlinx.coroutines.flow.collect
 
 
 class RegisterAgreeFragment :
