+++ app/src/main/java/com/surveasy/surveasy/presentation/intro/agree/RegisterTerm1.kt
@@ -1,22 +1,22 @@
-package com.surveasy.surveasy.presentation.register.agree
+package com.surveasy.surveasy.presentation.intro.agree
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.appcompat.widget.Toolbar
 import com.surveasy.surveasy.R
 
-class RegisterTerm2 : AppCompatActivity() {
+class RegisterTerm1 : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_register_term2)
+        setContentView(R.layout.activity_register_term1)
 
-        val term2Toolbar : Toolbar? = findViewById(R.id.ToolbarRegisterAgree2)
-        setSupportActionBar(term2Toolbar)
+        val term1Toolbar : Toolbar? = findViewById(R.id.ToolbarRegisterAgree1)
+        setSupportActionBar(term1Toolbar)
         if(supportActionBar != null){
             supportActionBar?.setDisplayHomeAsUpEnabled(true)
             supportActionBar?.setDisplayShowTitleEnabled(false)
         }
-        term2Toolbar?.setNavigationOnClickListener {
+        term1Toolbar?.setNavigationOnClickListener {
             onBackPressed()
         }
     }
