--- app/src/main/java/com/surveasy/surveasy/presentation/login/LoginActivity.kt
@@ -1,12 +1,0 @@
-package com.surveasy.surveasy.presentation.login
-
-import com.surveasy.surveasy.databinding.ActivityLoginBinding
-import com.surveasy.surveasy.presentation.base.BaseActivity
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class LoginActivity : BaseActivity<ActivityLoginBinding>(ActivityLoginBinding::inflate) {
-
-
-
-}
