+++ app/src/main/res/layout/activity_register_term1.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    tools:context=".presentation.register.agree.RegisterTerm1"
+    tools:context=".presentation.intro.agree.RegisterTerm1"
     android:background="@color/white"
     android:orientation="vertical">
