+++ app/src/main/res/layout/fragment_register_agree.xml
@@ -4,10 +4,10 @@
     xmlns:tools="http://schemas.android.com/tools">
 
     <data>
-        <import type="com.surveasy.surveasy.presentation.register.RegisterEventType"/>
+        <import type="com.surveasy.surveasy.presentation.intro.RegisterEventType"/>
         <variable
             name="vm"
-            type="com.surveasy.surveasy.presentation.register.RegisterViewModel" />
+            type="com.surveasy.surveasy.presentation.intro.RegisterViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
