+++ app/src/main/res/layout/fragment_register_input1.xml
@@ -4,11 +4,11 @@
 
     <data>
 
-        <import type="com.surveasy.surveasy.presentation.register.RegisterEventType" />
+        <import type="com.surveasy.surveasy.presentation.intro.RegisterEventType" />
 
         <variable
             name="vm"
-            type="com.surveasy.surveasy.presentation.register.RegisterViewModel" />
+            type="com.surveasy.surveasy.presentation.intro.RegisterViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
