+++ app/src/main/res/navigation/intro_nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/loginFragment"
-        android:name="com.surveasy.surveasy.presentation.register.LoginFragment"
+        android:name="com.surveasy.surveasy.presentation.intro.login.LoginFragment"
         android:label="LoginFragment" >
         <action
             android:id="@+id/action_loginFragment_to_existUserFragment"
@@ -18,11 +18,11 @@
     </fragment>
     <fragment
         android:id="@+id/existUserFragment"
-        android:name="com.surveasy.surveasy.presentation.register.ExistUserFragment"
+        android:name="com.surveasy.surveasy.presentation.intro.login.ExistUserFragment"
         android:label="ExistUserFragment" />
     <fragment
         android:id="@+id/registerAgreeFragment"
-        android:name="com.surveasy.surveasy.presentation.register.RegisterAgreeFragment"
+        android:name="com.surveasy.surveasy.presentation.intro.RegisterAgreeFragment"
         android:label="RegisterAgreeFragment"
         tools:layout="@layout/fragment_register_agree">
         <action
@@ -31,7 +31,7 @@
     </fragment>
     <fragment
         android:id="@+id/registerWarnFragment"
-        android:name="com.surveasy.surveasy.presentation.register.RegisterWarnFragment"
+        android:name="com.surveasy.surveasy.presentation.intro.RegisterWarnFragment"
         android:label="RegisterWarnFragment"
         tools:layout="@layout/fragment_register_warn">
         <action
@@ -40,7 +40,7 @@
     </fragment>
     <fragment
         android:id="@+id/registerInput1Fragment"
-        android:name="com.surveasy.surveasy.presentation.register.RegisterInput1Fragment"
+        android:name="com.surveasy.surveasy.presentation.intro.RegisterInput1Fragment"
         android:label="fragment_register_input1"
         tools:layout="@layout/fragment_register_input1" >
         <action
@@ -49,7 +49,7 @@
     </fragment>
     <fragment
         android:id="@+id/registerInput2Fragment"
-        android:name="com.surveasy.surveasy.presentation.register.RegisterInput2Fragment"
+        android:name="com.surveasy.surveasy.presentation.intro.RegisterInput2Fragment"
         android:label="fragment_register_input2"
         tools:layout="@layout/fragment_register_input2" />
 </navigation>
